Updated getImagesLib.js (added 3 lines, deleted 2 lines).

+++ getImagesLib.js
@@ -140,12 +140,13 @@ function compositeDates(images,composite,bandNames){
   if(bandNames === null || bandNames === undefined){
      bandNames = ee.Image(images.first()).bandNames();
   }else{images = images.select(bandNames);composite = composite.select(bandNames)}
-
+  
+  composite = composite.float();
   var bns = ee.Image(images.first()).bandNames().map(function(bn){return ee.String(bn).cat('_diff')});
 
   //Function to get the abs diff from a given composite *-1
   function getDiff(img,composite){
-    var out = img.subtract(composite).abs().multiply(-1).rename(bns);
+    var out = img.float().subtract(composite).abs().multiply(-1).rename(bns);
     return img.addBands(out);
   }
